backend/tests/requirements-dev.txt
# Development dependencies for testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
azure-functions==1.18.0
azure-cosmos==4.5.1
